Add container folder and move App.js in it
